cleanup getPreviousPlayerScoreFromId
This one is interesting. Near as I can tell:
1) It is incorrectly just checking for week 1 stats, not the week thatis currently being asked for
2) It iterates the player collection for no real reason, and then only uses the passed in player_id to look for something in previousStats anyway.

So I'm removing the iteration of the player array, and making it check if the week in question exists in the array.

I could imagine that iterating the player array was to ensure a valid player_id, but as we are calling this only from iterating the player array I feel confident it'll always be good.